AnalizX/requirements.txt
pandas
numpy
streamlit
plotly
Pillow
